+++ src/mcp_tools.ts
@@ -1,4 +1,34 @@
 import { OpenAITool } from "./type";
+import { toolsBase } from "./tools_base";
+
+export const tools = toolsBase.map((tool: OpenAITool) => {
+  const { type, function: func } = tool;
+  const { name, description, parameters } = func;
+  return {
+    type,
+    function: {
+      name: `create${name}Slide`,
+      description: `Create or update ${description}`,
+      parameters,
+    },
+  };
+});
+
+export const toolsForBeat = toolsBase.map((tool: OpenAITool) => {
+  const { type, function: func } = tool;
+  const { name, description, parameters } = func;
+  return {
+    type,
+    function: {
+      name: `updateBeatStyleTo${name}`,
+      description: `Change the beat image to ${description}`,
+      parameters,
+    },
+  };
+});
+
+// updateBeatImageForXXStyle
+// Change the beat image to
 
 export const mcp_tools: OpenAITool[] = [
   {
